requirements.txt
Pillow==9.4.0
Random_Word==1.0.11
